+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Make ResponseValidator errors easier to read
+
 # 0.6.0
 
 - Set the content-type based on the OpenAPI description [#29](https://github.com/ahx/openapi-first/pull/29)